+++ src/main/java/org/mufuku/yaoocai/v1/compiler/translator/IfJumpTable.java
@@ -8,23 +8,23 @@ import java.util.List;
  */
 class IfJumpTable {
 
-    private final List<Short> blockSizes = new ArrayList<>();
+    private final List<Byte> blockSizes = new ArrayList<>();
 
-    private final List<Short> expressionSizes = new ArrayList<>();
+    private final List<Byte> expressionSizes = new ArrayList<>();
 
-    void addEntry(short blockSize, short expressionSize) {
+    void addEntry(byte blockSize, byte expressionSize) {
         this.blockSizes.add(blockSize);
         this.expressionSizes.add(expressionSize);
     }
 
-    short getIfJumpOffset(int index) {
-        short jumpOffset = 1;
+    byte getIfJumpOffset(int index) {
+        byte jumpOffset = 1;
         jumpOffset += blockSizes.get(index);
         return jumpOffset;
     }
 
-    short getEndJumpOffset(int index) {
-        short jumpOffset = 1;
+    byte getEndJumpOffset(int index) {
+        byte jumpOffset = 1;
         for (int j = index + 1; j < blockSizes.size(); j++) {
             jumpOffset += blockSizes.get(j);
             jumpOffset += expressionSizes.get(j);
