+++ src/main/java/org/mufuku/yaoocai/v1/vm/LocalVariableStack.java
@@ -7,13 +7,13 @@ import java.util.Map;
  * @author Andreas Etzlstorfer (a.etzlstorfer@gmail.com)
  */
 class LocalVariableStack {
-    private final Map<Short, Object> variables = new HashMap<>();
+    private final Map<Byte, Object> variables = new HashMap<>();
 
-    void setValue(short index, Object value) {
+    void setValue(byte index, Object value) {
         variables.put(index, value);
     }
 
-    Object getValue(short index) {
+    Object getValue(byte index) {
         return variables.get(index);
     }
 }
